+++ pkg/cmd/time-entry/report/util/report.go
@@ -52,7 +52,9 @@ func (rf ReportFlags) Check() error {
 // NewReportFlags helps creating a util.ReportFlags for report commands
 func NewReportFlags() ReportFlags {
 	return ReportFlags{
-		OutputFlags: util.OutputFlags{TimeFormat: timehlp.FullTimeFormat},
+		OutputFlags: util.OutputFlags{
+			TimeFormat: timehlp.FullTimeFormat,
+		},
 	}
 }